# docs/reference/openapi.yaml
openapi: 3.1.0
info:
  title: Yantra4D API
  version: 1.0.0
  description: |
    REST API for the Yantra4D parametric 3D design platform.
    Provides OpenSCAD rendering, project management, AI-assisted design,
    Git operations, and analytics.
  license:
    name: AGPL-3.0
    url: https://opensource.org/license/agpl-v3/
servers:
  - url: http://localhost:5000
    description: Local development
  - url: https://api.yantra4d.com
    description: Production
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Janua-issued JWT token
...
paths:
...
  /api/projects/{slug}/download/stl/{filename}:
    get:
      tags: [download]
      summary: Download STL file
      operationId: downloadStl
      security:
        - bearerAuth: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: filename
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: STL binary
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "401":
          description: Authentication required
        "404":
          description: File not found
  /api/projects/{slug}/download/scad/{filename}:
    get:
      tags: [download]
      summary: Download SCAD source file
      operationId: downloadScad
      security:
        - bearerAuth: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: filename
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: SCAD source file
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "403":
          description: File not in manifest allowlist
        "404":
          description: File not found

  # ── Editor ──────────────────────────────────────────────
...